Algebra.Matrix.Generic.Mutable

Documentation

class (Vector (Column mat) a, Vector (Row mat) a) => MMatrix mat a where

Mutable, row-based 0-origin matrix

Minimal complete definition

basicUnsafeNew, basicInitialise, basicRowCount, basicColumnCount, unsafeGetRow, unsafeGetColumn, unsafeFromRows, unsafeCopy, unsafeRead, unsafeWrite, basicSet, basicUnsafeIMapRowM, unsafeScaleRow

Methods

basicUnsafeNew :: PrimMonad m => Size -> Size -> m (mat (PrimState m) a)

basicUnsafeNew n m creates a mutable matrix with n rows and m columns, without initialisation. This method should not be used directly, use unsafeNew instead.

basicInitialise :: PrimMonad m => mat (PrimState m) a -> m ()

basicRowCount :: mat s a -> Size

basicColumnCount :: mat s a -> Size

unsafeGetRow :: PrimMonad m => Index -> mat (PrimState m) a -> m (Row mat a)

unsafeGetColumn :: PrimMonad m => Index -> mat (PrimState m) a -> m (Column mat a)

unsafeFill :: PrimMonad m => Size -> Size -> a -> m (mat (PrimState m) a)

unsafeFromRow :: PrimMonad m => Row mat a -> m (mat (PrimState m) a)

Construct a mutable matrix consisting of a single row, perhaps without any copy.

unsafeFromRows :: PrimMonad m => [Row mat a] -> m (mat (PrimState m) a)

Construct a mutable matrix consisting a single column, perhaps without any copy.

unsafeFromColumn :: PrimMonad m => Column mat a -> m (mat (PrimState m) a)

Construct a mutable matrix consisting a single column, perhaps without any copy.

unsafeFromColumns :: PrimMonad m => [Column mat a] -> m (mat (PrimState m) a)

unsafeCopy :: PrimMonad m => mat (PrimState m) a -> mat (PrimState m) a -> m ()

usnafeCopy target source copies the content of source to target, without boundary check.

unsafeRead :: PrimMonad m => mat (PrimState m) a -> Index -> Index -> m a

unsafeRead i j m reads the value at ith row in jth column in m, without boundary check.

N.B. Rows and columns are regarded as zero-origin, not 1!.

unsafeWrite :: PrimMonad m => mat (PrimState m) a -> Index -> Index -> a -> m ()

basicSet :: PrimMonad m => mat (PrimState m) a -> a -> m ()

basicUnsafeIMapRowM :: PrimMonad m => mat (PrimState m) a -> Index -> (Index -> a -> m a) -> m ()

basicUnsafeIMapRow :: PrimMonad m => mat (PrimState m) a -> Index -> (Index -> a -> a) -> m ()

basicUnsafeSwapRows :: PrimMonad m => mat (PrimState m) a -> Index -> Index -> m ()

unsafeScaleRow :: (PrimMonad m, Commutative a) => mat (PrimState m) a -> Index -> a -> m ()

unsafeGenerate :: PrimMonad m => Size -> Size -> (Index -> Index -> a) -> m (mat (PrimState m) a)

unsafeGenerateM :: PrimMonad m => Size -> Size -> (Index -> Index -> m a) -> m (mat (PrimState m) a)

toRows :: PrimMonad m => mat (PrimState m) a -> m [Row mat a]

toColumns :: PrimMonad m => mat (PrimState m) a -> m [Column mat a]

type Index = Int

type Size = Int

type family Column (mat :: k) :: Type -> Type

type family Row (mat :: k) :: Type -> Type

new :: (MMatrix mat a, PrimMonad m) => Size -> Size -> m (mat (PrimState m) a)

new n m creates a mutable matrix with n rows and m columns.

unsafeNew :: (MMatrix mat a, PrimMonad m) => Size -> Size -> m (mat (PrimState m) a)

unsafeNew n m creates a mutable matrix with n rows and m columns, without memory initialisation.

See also: new.

copy :: (PrimMonad m, MMatrix mat a) => mat (PrimState m) a -> mat (PrimState m) a -> m ()

clone :: (PrimMonad m, MMatrix mat a) => mat (PrimState m) a -> m (mat (PrimState m) a)

generate :: (PrimMonad m, MMatrix mat a) => Int -> Int -> (Index -> Index -> a) -> m (mat (PrimState m) a)

generateM :: (PrimMonad m, MMatrix mat a) => Int -> Int -> (Index -> Index -> m a) -> m (mat (PrimState m) a)

fromRow :: (PrimMonad m, MMatrix mat a) => Row mat a -> m (mat (PrimState m) a)

fromColumn :: (PrimMonad m, MMatrix mat a) => Column mat a -> m (mat (PrimState m) a)

getRow :: (MMatrix mat a, PrimMonad m) => Index -> mat (PrimState m) a -> m (Row mat a)

getRow n mat retrieves nth row in mat

N.B. Index is considered as 0-origin, NOT 1!

getColumn :: (MMatrix mat a, PrimMonad m) => Index -> mat (PrimState m) a -> m (Column mat a)

getColumn n mat retrieves nth colun in mat

N.B. Index is considered as 0-origin, NOT 1!

imapRow :: (PrimMonad m, MMatrix mat a) => Index -> (Index -> a -> a) -> mat (PrimState m) a -> m ()

imapRow i f m mutates ith row in the matrix m by applying f with column index.

See also: mapRow, imapColumn, mapColumn.

N.B. Index is considered as 0-origin, NOT 1!

mapRow :: (PrimMonad m, MMatrix mat a) => Index -> (a -> a) -> mat (PrimState m) a -> m ()

mapRow i f m mutates ith row in the matrix m by applying f.

See also: imapRow, imapColumn, mapColumn.

N.B. Index is considered as 0-origin, NOT 1!

fill :: (PrimMonad m, MMatrix mat a) => Size -> Size -> a -> m (mat (PrimState m) a)

fill n m c creates a mutable constant matrix with n rows and m columns.

N.B. Index is considered as 0-origin, NOT 1!

read :: (PrimMonad m, MMatrix mat a) => Index -> Index -> mat (PrimState m) a -> m a

read i j m reads the value at ith row in jth column in m

N.B. Index is considered as 0-origin, NOT 1!

write :: (PrimMonad m, MMatrix mat a) => Index -> Index -> mat (PrimState m) a -> a -> m ()

read i j m writes the value at ith row in jth column in m

N.B. Index is considered as 0-origin, NOT 1!

scaleRow :: (Multiplicative a, Commutative a, MMatrix mat a, PrimMonad m) => Index -> a -> mat (PrimState m) a -> m ()

scaleRowL i c m multiplies every element in ith row in m by c, from right.

See also: scaleRowL and scaleRow

unsafeSwapRows :: (PrimMonad m, MMatrix mat a) => Index -> Index -> mat (PrimState m) a -> m ()

unsafeSwapRows n m mat swaps nth and mth rows in m, without boundary check.

See also: swapRows.

swapRows :: (PrimMonad m, MMatrix mat a) => Index -> Index -> mat (PrimState m) a -> m ()

swapRows n m mat swaps nth and mth rows in m.

See also: unsafeSwapRows.

combineRows :: (Semiring a, Commutative a, PrimMonad m, MMatrix mat a) => Index -> a -> Index -> mat (PrimState m) a -> m ()

combineRows i c j mat adds scalar multiple of jth row by c to ith.

gaussReduction :: (Eq a, PrimMonad m, Field a, Normed a, MMatrix mat a) => mat (PrimState m) a -> m (mat (PrimState m) a, mat (PrimState m) a, a)

Performs Gaussian reduction to given matrix, returns the pair of triangulated matrix, pivot matrix and determinant.

unsafeGaussReduction :: (MMatrix mat a, Normed a, Eq a, Field a, PrimMonad m) => mat (PrimState m) a -> m (mat (PrimState m) a, a)

Performs in-place gaussian reduction to the mutable matrix, and returns the pivoting matrix and determinant.